feat(bin): add users_ext binary for fetching user details
Adds a new binary `users_ext` to concurrently fetch detailed information for a large, predefined list of user IDs. This tool is useful for bulk data extraction.

This commit also includes a critical fix for the pagination logic. The loop condition was incorrect, preventing pagination from working as intended. The initial total page count is now set to `u64::MAX` to ensure the first page is always fetched.

Additionally, the `page` field in API requests has been changed from `u16` to `usize` to better accommodate large result sets and align with standard Rust types for indexing.

+++ libft-api/src/api/user/users_id.rs
@@ -9,7 +9,7 @@ pub struct FtApiUsersIdRequest {
     pub sort: Option<Vec<FtSortOption>>,
     pub range: Option<Vec<FtRangeOption>>,
     pub filter: Option<Vec<FtFilterOption>>,
-    pub page: Option<u16>,
+    pub page: Option<usize>,
     pub per_page: Option<u8>,
 }
